[vuihandler]: Fix int64_t in int32_t in vuihandler FE header

+++ ME/base/so3/soo/include/soo/dev/vuihandler.h
@@ -43,7 +43,7 @@
 #define VUIHANDLER_PERIOD	1000
 
 typedef struct {
-	int64_t		slotID;
+	int32_t		slotID;
 	uint8_t		type;
 	uint8_t		payload[0];
 } vuihandler_pkt_t;
